+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Changelog
 
+### Version 0.5.1
+* Tweak behavior of optional metric name sanitizer; pypy support
+
+### Version 0.5.0
+* Release 0.5.0 - adds the option to sanitize metric names; other minor changes
+
+### Version 0.4.14
+* Fix issues in Gauge#add and Counter#add per #69
+
 ### Version 0.4.13
 * Update setup.py to include supported Python versions